+++ census/exceptions.py
@@ -1,0 +1,8 @@
+"""
+exceptions.py
+=================
+census exceptions
+"""
+
+class CensusException(Exception):
+    pass