+++ get_census/exceptions.py
@@ -1,6 +1,0 @@
-"""
-get_census exceptions
-"""
-
-class GetCensusException(Exception):
-    pass