+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added 
 
+- glossary
+- show_api_keys
 - public folder to hold webpage documents for gitlab.
 - docs folder to hold webpage documents for github.
 - Set up environment page. 
@@ -11,6 +13,8 @@
 
 ### Changed
 
+- get_census --> census
+- Reorganized the imports
 - docs/source --> docsource
 - Moved Makefiles into docsource folder.
 - Makefile: added github and gitlab actions (use `make gitlab`, and `make github`)