+++ docsource/CHANGELOG.md
@@ -4,12 +4,17 @@
 
 ### Added 
 
+- glossary
+- show_api_keys
 - public folder to hold webpage documents for gitlab.
 - docs folder to hold webpage documents for github.
+- Set up environment page. 
 
 
 ### Changed
 
+- get_census --> census
+- Reorganized the imports
 - docs/source --> docsource
 - Moved Makefiles into docsource folder.
 - Makefile: added github and gitlab actions (use `make gitlab`, and `make github`)